refactor: break it out
